requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
elasticsearch>=5.0.0,<6.0.0
schematics == 1.1.2
google-cloud-vision
google-cloud-logging
git+https://github.com/ciandt-d1/cvtool-image-hashes-client.git
google-cloud-storage
google-cloud-bigquery
flask-cors
cryptography == 2.0.3